:art: refactor : 한 노선의 역을 지우는 기능의 클래스와 메서드명을 더 명확하게 개선

+++ src/main/java/subway/path/domain/Path.java
@@ -78,7 +78,7 @@ public class Path {
             return 0;
         }
         
-        return (long) ((Math.ceil((distance - 1) / increase) + 1) * 100);
+        return (long) ((Math.ceil((double) (distance - 1) / increase) + 1) * 100);
     }
     
     private boolean isDistanceNegative(final long distance) {
